+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Foreign Chain
+- `Registry` contract has additional method `atomicUpdate`
+
+### Changed
+- redeployment process for `ForeignChain` is different from `Chain`
 
 ## [3.2.0] - 2021-08-31
 ### Changed